Refresh federated geometry on immutable mesh replacement
