+++ .changeset/federated-mesh-replacement.md
@@ -1,0 +1,5 @@
+---
+"@ifc-lite/viewer": patch
+---
+
+Refresh federated viewport geometry when a command replaces a model's mesh array, including replacements that restore additional objects. Keep same-array streaming appends incremental.
